+++ catalog/resource_workspace_binding_test.go
@@ -7,13 +7,13 @@ import (
 	"github.com/databricks/terraform-provider-databricks/qa"
 )
 
-func TestCatalogWorkspaceBindingsCornerCases(t *testing.T) {
-	qa.ResourceCornerCases(t, ResourceCatalogWorkspaceBinding(),
+func TestWorkspaceBindingsCornerCases(t *testing.T) {
+	qa.ResourceCornerCases(t, ResourceWorkspaceBinding(),
 		qa.CornerCaseID("1234567890101112|catalog|my_catalog"),
 		qa.CornerCaseSkipCRUD("create"))
 }
 
-func TestCatalogWorkspaceBindings_Create(t *testing.T) {
+func TestWorkspaceBindings_Create(t *testing.T) {
 	qa.ResourceFixture{
 		Fixtures: []qa.HTTPFixture{
 			{
@@ -50,7 +50,7 @@ func TestCatalogWorkspaceBindings_Create(t *testing.T) {
 				},
 			},
 		},
-		Resource: ResourceCatalogWorkspaceBinding(),
+		Resource: ResourceWorkspaceBinding(),
 		Create:   true,
 		HCL: `
 		catalog_name = "my_catalog"
@@ -59,7 +59,7 @@ func TestCatalogWorkspaceBindings_Create(t *testing.T) {
 	}.ApplyNoError(t)
 }
 
-func TestCatalogWorkspaceBindingsReadOnly_Create(t *testing.T) {
+func TestWorkspaceBindingsReadOnly_Create(t *testing.T) {
 	qa.ResourceFixture{
 		Fixtures: []qa.HTTPFixture{
 			{
@@ -96,7 +96,7 @@ func TestCatalogWorkspaceBindingsReadOnly_Create(t *testing.T) {
 				},
 			},
 		},
-		Resource: ResourceCatalogWorkspaceBinding(),
+		Resource: ResourceWorkspaceBinding(),
 		Create:   true,
 		HCL: `
 		catalog_name = "my_catalog"
@@ -106,7 +106,104 @@ func TestCatalogWorkspaceBindingsReadOnly_Create(t *testing.T) {
 	}.ApplyNoError(t)
 }
 
-func TestCatalogWorkspaceBindingsReadImport(t *testing.T) {
+func TestSecurableWorkspaceBindings_Create(t *testing.T) {
+	qa.ResourceFixture{
+		Fixtures: []qa.HTTPFixture{
+			{
+				Method:   "PATCH",
+				Resource: "/api/2.1/unity-catalog/bindings/catalog/my_catalog",
+				ExpectedRequest: catalog.UpdateWorkspaceBindingsParameters{
+					Add: []catalog.WorkspaceBinding{
+						{
+							BindingType: catalog.WorkspaceBindingBindingTypeBindingTypeReadOnly,
+							WorkspaceId: int64(1234567890101112),
+						},
+					},
+					SecurableName: "my_catalog",
+					SecurableType: "catalog",
+				},
+				Response: catalog.WorkspaceBindingsResponse{
+					Bindings: []catalog.WorkspaceBinding{
+						{
+							BindingType: catalog.WorkspaceBindingBindingTypeBindingTypeReadOnly,
+							WorkspaceId: int64(1234567890101112),
+						},
+					},
+				},
+			}, {
+				Method:   "GET",
+				Resource: "/api/2.1/unity-catalog/bindings/catalog/my_catalog?",
+				Response: catalog.WorkspaceBindingsResponse{
+					Bindings: []catalog.WorkspaceBinding{
+						{
+							BindingType: catalog.WorkspaceBindingBindingTypeBindingTypeReadOnly,
+							WorkspaceId: int64(1234567890101112),
+						},
+					},
+				},
+			},
+		},
+		Resource: ResourceWorkspaceBinding(),
+		Create:   true,
+		HCL: `
+		securable_name = "my_catalog"
+		securable_type = "catalog"
+		workspace_id   = "1234567890101112"
+		binding_type   = "BINDING_TYPE_READ_ONLY"
+		`,
+	}.ApplyNoError(t)
+}
+
+func TestSecurableWorkspaceBindings_Delete(t *testing.T) {
+	qa.ResourceFixture{
+		Fixtures: []qa.HTTPFixture{
+			{
+				Method:   "PATCH",
+				Resource: "/api/2.1/unity-catalog/bindings/catalog/my_catalog",
+				ExpectedRequest: catalog.UpdateWorkspaceBindingsParameters{
+					Remove: []catalog.WorkspaceBinding{
+						{
+							BindingType: catalog.WorkspaceBindingBindingTypeBindingTypeReadOnly,
+							WorkspaceId: int64(1234567890101112),
+						},
+					},
+					SecurableName: "my_catalog",
+					SecurableType: "catalog",
+				},
+				Response: catalog.WorkspaceBindingsResponse{
+					Bindings: []catalog.WorkspaceBinding{
+						{
+							BindingType: catalog.WorkspaceBindingBindingTypeBindingTypeReadOnly,
+							WorkspaceId: int64(1234567890101112),
+						},
+					},
+				},
+			}, {
+				Method:   "GET",
+				Resource: "/api/2.1/unity-catalog/bindings/catalog/my_catalog?",
+				Response: catalog.WorkspaceBindingsResponse{
+					Bindings: []catalog.WorkspaceBinding{
+						{
+							BindingType: catalog.WorkspaceBindingBindingTypeBindingTypeReadOnly,
+							WorkspaceId: int64(1234567890101112),
+						},
+					},
+				},
+			},
+		},
+		Resource: ResourceWorkspaceBinding(),
+		Delete:   true,
+		ID:       "1234567890101112|catalog|my_catalog",
+		HCL: `
+		securable_name = "my_catalog"
+		securable_type = "catalog"
+		workspace_id   = "1234567890101112"
+		binding_type   = "BINDING_TYPE_READ_ONLY"
+		`,
+	}.ApplyNoError(t)
+}
+
+func TestWorkspaceBindingsReadImport(t *testing.T) {
 	qa.ResourceFixture{
 		Fixtures: []qa.HTTPFixture{
 			{
@@ -122,7 +219,7 @@ func TestCatalogWorkspaceBindingsReadImport(t *testing.T) {
 				},
 			},
 		},
-		Resource: ResourceCatalogWorkspaceBinding(),
+		Resource: ResourceWorkspaceBinding(),
 		ID:       "1234567890101112|catalog|my_catalog",
 		New:      true,
 		Read:     true,
@@ -133,16 +230,16 @@ func TestCatalogWorkspaceBindingsReadImport(t *testing.T) {
 	})
 }
 
-func TestCatalogWorkspaceBindingsReadErrors(t *testing.T) {
+func TestWorkspaceBindingsReadErrors(t *testing.T) {
 	qa.ResourceFixture{
-		Resource: ResourceCatalogWorkspaceBinding(),
+		Resource: ResourceWorkspaceBinding(),
 		ID:       "1234567890101112|catalog",
 		New:      true,
 		Read:     true,
 	}.ExpectError(t, "incorrect binding id: 1234567890101112|catalog. Correct format: <workspace_id>|<securable_type>|<securable_name>")
 
 	qa.ResourceFixture{
-		Resource: ResourceCatalogWorkspaceBinding(),
+		Resource: ResourceWorkspaceBinding(),
 		ID:       "A234567890101112|catalog|my_catalog",
 		New:      true,
 		Read:     true,
